+++ lib/src/buttons/radio_button.dart
@@ -37,7 +37,7 @@ class RadioButton extends StatelessWidget {
   final double size;
 
   /// The color of the border when [value] is true. If null,
-  /// [Style.primaryColor] is used
+  /// [MacosThemeData.primaryColor] is used
   final Color? onColor;
 
   /// The color of the border when [value] is false.
@@ -94,7 +94,7 @@ class RadioButton extends StatelessWidget {
             border: Border.all(
               style: isDisabled ? BorderStyle.none : BorderStyle.solid,
               width: value ? size / 4.0 : 1,
-              color: DynamicColorX.macosResolve(
+              color: MacosDynamicColor.resolve(
                 value ? onColor ?? theme.primaryColor : offColor,
                 context,
               ),
@@ -106,7 +106,7 @@ class RadioButton extends StatelessWidget {
           child: Container(
             decoration: BoxDecoration(
               shape: BoxShape.circle,
-              color: DynamicColorX.macosResolve(
+              color: MacosDynamicColor.resolve(
                 innerColor ??
                     (isDisabled
                         ? CupertinoColors.quaternarySystemFill