+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [0.1.2]
 * Updated the theme api
+  * Properties in `MacosThemeData` can't be null
+  * Renamed `DynamicColorX` to `MacosDynamicColor`
+  * Added the method `lerp` on all theme datas.
 
 ## [0.1.1]
 * Implemented `Label` ([#61](https://github.com/GroovinChip/macos_ui/issues/61))